Fix issues with empty or invalid patterns and other fixes.
Don't error when listing entries without a pattern.

Don't try to regex match a pattern that doesn't exist.

Make sure the pattern compiles as a regex at the time the entry is added.

If no mask is passed set it to *!*@*.

If nick is passed as mask change it to nick!*@*.